Initial commit
This shell script will be able to change the registered IP of a running masternode on the same network when it detects the IP has changed.  It should be run periodically from a cron job or timer and needs access to another dashd wallet with a small amount of Dash in it to pay for the TX fee.